Refactoring uniforms, attributes, and scopes out of mesh to generic gpu module.
