+++ lib/core/src/display/symbol.rs
@@ -18,10 +18,10 @@ use crate::data::function::callback::*;
 use crate::debug::stats::Stats;
 use crate::display::render::webgl::Context;
 use crate::display::render::webgl;
-use crate::display::symbol::geometry::primitive::mesh::buffer::IsBuffer;
-use crate::display::symbol::geometry::primitive::mesh::scope::uniform::AnyUniform;
-use crate::display::symbol::geometry::primitive::mesh::scope::uniform::AnyUniformOps;
-use crate::display::symbol::geometry::primitive::mesh::scope::uniform::UniformScope;
+use crate::system::gpu::buffer::IsBuffer;
+use crate::system::gpu::scope::uniform::AnyUniform;
+use crate::system::gpu::scope::uniform::AnyUniformOps;
+use crate::system::gpu::scope::uniform::UniformScope;
 use crate::display::symbol::geometry::primitive::mesh;
 use crate::promote;
 use crate::promote_all;
