+++ lib/core/src/display/symbol/registry.rs
@@ -10,12 +10,12 @@ use crate::data::opt_vec::OptVec;
 use crate::debug::stats::Stats;
 use crate::display::camera::Camera2D;
 use crate::display::render::webgl::Context;
-use crate::display::symbol::geometry::primitive::mesh::scope::uniform::UniformScope;
-use crate::display::symbol::geometry::primitive::mesh::scope::uniform::Uniform;
 use crate::display::symbol;
 use crate::promote;
 use crate::promote_all;
 use crate::promote_symbol_types;
+use crate::system::gpu::scope::uniform::Uniform;
+use crate::system::gpu::scope::uniform::UniformScope;
 use crate::system::web::group;
 use crate::system::web::Logger;
 use eval_tt::*;
