+++ lib/core/src/system/gpu.rs
@@ -1,1 +1,3 @@
+pub mod buffer;
 pub mod data;
+pub mod scope;
