+++ lib/core/src/system/gpu/buffer.rs
@@ -1,7 +1,5 @@
 #![allow(missing_docs)]
 
-#[warn(missing_docs)]
-
 use crate::prelude::*;
 
 use crate::closure;
