+++ lib/core/src/system/gpu/scope.rs
