+++ lib/core/src/system/gpu/scope/uniform.rs
